Moved old deprecation module to eva-accession-dbsnp-deprecate
