+++ eva-accession-dbsnp-deprecate/src/main/java/uk/ac/ebi/eva/accession/dbsnp/deprecate/configuration/BeanNames.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2019 EMBL - European Bioinformatics Institute
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.ebi.eva.accession.dbsnp.deprecate.configuration;
+
+public class BeanNames {
+
+    public static final String DEPRECABLE_CLUSTERED_VARIANTS_READER = "DEPRECABLE_CLUSTERED_VARIANTS_READER";
+
+    public static final String DEPRECATION_WRITER = "DEPRECATION_WRITER";
+
+    public static final String DEPRECATE_CLUSTERED_VARIANTS_STEP = "DEPRECATE_CLUSTERED_VARIANTS_STEP";
+
+    public static final String DEPRECATE_CLUSTERED_VARIANTS_JOB = "DEPRECATE_CLUSTERED_VARIANTS_JOB";
+
+    public static final String DEPRECATION_PROGRESS_LISTENER = "DEPRECATION_PROGRESS_LISTENER";
+
+}
