+++ eva-accession-dbsnp-deprecate/src/main/java/uk/ac/ebi/eva/accession/dbsnp/deprecate/configuration/InputParametersConfiguration.java
@@ -1,0 +1,17 @@
+package uk.ac.ebi.eva.accession.dbsnp.deprecate.configuration;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import uk.ac.ebi.eva.accession.dbsnp.deprecate.parameters.InputParameters;
+
+@Configuration
+public class InputParametersConfiguration {
+
+    @Bean
+    @ConfigurationProperties(prefix = "parameters")
+    public InputParameters inputParameters() {
+        return new InputParameters();
+    }
+}
