+++ eva-accession-dbsnp-deprecate/src/main/resources/application.properties
