+++ eva-accession-dbsnp-deprecate/src/test/resources/application.properties
