+++ eva-accession-dbsnp-deprecate/src/test/resources/test-data/dbsnpClusteredVariantEntityDeclustered.json
