+++ eva-accession-dbsnp-deprecate/src/test/resources/test-data/dbsnpSubmittedVariantEntity.json
