--- eva-accession-deprecate/src/main/java/uk/ac/ebi/eva/accession/deprecate/parameters/InputParameters.java
@@ -1,51 +1,0 @@
-/*
- * Copyright 2019 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.ac.ebi.eva.accession.deprecate.parameters;
-
-import java.util.List;
-
-public class InputParameters {
-
-    private int chunkSize;
-
-    private List<String> assemblyAccession;
-
-    private boolean deprecateAll;
-
-    public int getChunkSize() {
-        return chunkSize;
-    }
-
-    public void setChunkSize(int chunkSize) {
-        this.chunkSize = chunkSize;
-    }
-
-    public List<String> getAssemblyAccession() {
-        return assemblyAccession;
-    }
-
-    public void setAssemblyAccession(List<String> assemblyAccession) {
-        this.assemblyAccession = assemblyAccession;
-    }
-
-    public Boolean getDeprecateAll() {
-        return deprecateAll;
-    }
-
-    public void setDeprecateAll(Boolean deprecateAll) {
-        this.deprecateAll = deprecateAll;
-    }
-}
